refactor: improve namespace handling
Previously the search() method used a wildcard namespace allowing all namespaces to be included in search results, but that added messy results like files or images which were not desired.

Now the search() method uses the previously fetched (by the API.site_init() method) namespaces to filter out the results.